hierarchy: refactor
