+++ passes/hierarchy/util/Makefile.inc
@@ -1,0 +1,7 @@
+OBJS += passes/hierarchy/util/clean.o
+OBJS += passes/hierarchy/util/generate.o
+OBJS += passes/hierarchy/util/interfaces.o
+OBJS += passes/hierarchy/util/misc.o
+OBJS += passes/hierarchy/util/positionals.o
+OBJS += passes/hierarchy/util/top.o
+OBJS += passes/hierarchy/util/verilog.o
