+++ passes/hierarchy/util/generate.h
@@ -1,0 +1,44 @@
+/*
+*  yosys -- Yosys Open SYnthesis Suite
+*
+*  Copyright (C) 2012  Claire Xenia Wolf <claire@yosyshq.com>
+*  Copyright (C) 2018  Ruben Undheim <ruben.undheim@gmail.com>
+*
+*  Permission to use, copy, modify, and/or distribute this software for any
+*  purpose with or without fee is hereby granted, provided that the above
+*  copyright notice and this permission notice appear in all copies.
+*
+*  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+*  WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+*  MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+*  ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+*  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+*  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+*  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*
+*/
+
+#ifndef HIERARCHY_GENERATE_H
+#define HIERARCHY_GENERATE_H
+
+#include "kernel/yosys.h"
+
+YOSYS_NAMESPACE_BEGIN
+
+namespace Hierarchy {
+	struct generate_port_decl_t {
+		bool input, output;
+		string portname;
+		int index;
+	};
+	struct Generator {
+		std::vector<std::string> cells;
+		std::vector<generate_port_decl_t> ports;
+		void parse_arg(const char* p);
+		void generate(Design* design);
+	};
+};
+
+YOSYS_NAMESPACE_END
+
+#endif /* HIERARCHY_GENERATE_H */
