+++ passes/hierarchy/util/top.h
@@ -1,0 +1,49 @@
+/*
+ *  yosys -- Yosys Open SYnthesis Suite
+ *
+ *  Copyright (C) 2012  Claire Xenia Wolf <claire@yosyshq.com>
+ *  Copyright (C) 2018  Ruben Undheim <ruben.undheim@gmail.com>
+ *
+ *  Permission to use, copy, modify, and/or distribute this software for any
+ *  purpose with or without fee is hereby granted, provided that the above
+ *  copyright notice and this permission notice appear in all copies.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ *  WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ *  MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ *  ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ *  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ *  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ *  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ *
+ */
+
+#ifndef HIERARCHY_TOP_H
+#define HIERARCHY_TOP_H
+
+#include "kernel/yosys.h"
+
+YOSYS_NAMESPACE_BEGIN
+
+namespace Hierarchy {
+    struct TopModulePrepare {
+        Design* design;
+        const std::map<std::string, std::string>& params;
+        bool is_from_verific = false;
+        Module* load_top(const string& name, bool auto_top);
+    private:
+        Module* load_top_module_regular(const string& name);
+        Module* load_top_module_verific(bool& is_from_verific, string name, Module* top_mod);
+        void derive_all();
+        Module* load_top_module_attribute();
+        Module* load_top_module_auto();
+        Module* rederived_top(Module* top_mod);
+        void clear_initial_top_attribute(Module* top_mod);
+    };
+
+    void ensure_unique_top_attribute(Module* top_mod, Design* design);
+};
+
+YOSYS_NAMESPACE_END
+
+#endif /* HIERARCHY_TOP_H */
